refactor(shorebird_cli): create `Gradlew`
