+++ packages/shorebird_cli/bin/shorebird.dart
@@ -10,6 +10,7 @@ import 'package:shorebird_cli/src/cache.dart';
 import 'package:shorebird_cli/src/code_push_client_wrapper.dart';
 import 'package:shorebird_cli/src/command_runner.dart';
 import 'package:shorebird_cli/src/doctor.dart';
+import 'package:shorebird_cli/src/gradlew.dart';
 import 'package:shorebird_cli/src/ios_deploy.dart';
 import 'package:shorebird_cli/src/java.dart';
 import 'package:shorebird_cli/src/logger.dart';
@@ -31,6 +32,7 @@ Future<void> main(List<String> args) async {
         codePushClientWrapperRef,
         doctorRef,
         engineConfigRef,
+        gradlewRef,
         iosDeployRef,
         javaRef,
         loggerRef,
